+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppstier/domain/TierCalculation.kt
@@ -3,6 +3,7 @@ package uk.gov.justice.digital.hmpps.hmppstier.domain
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Component
 import uk.gov.justice.digital.hmpps.hmppstier.domain.enums.*
+import java.math.BigDecimal
 
 @Component
 class TierCalculation {
@@ -79,7 +80,7 @@ class TierCalculation {
       }
     }
 
-    private fun getRsrPoints(rsrScore: Int?): Int {
+    private fun getRsrPoints(rsrScore: BigDecimal?): Int {
       return when {
         rsrScore != null && rsrScore >= RsrThresholds.TIER_B_RSR.num -> {
           criteria.add(TierMatchCriteria.RSR_IN_TIER_B)
